requirements.txt
cmake==3.18.4
dlib==19.22.0
scipy==1.6.2
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
face-recognition==1.3.0
